Migrated collaborators to chat view.
