+++ src/components/concepts/MarkupHTMLView.svelte
@@ -11,34 +11,38 @@
     interface Props {
         markup: Markup | string[] | string;
         inline?: boolean;
+        note?: boolean;
     }
 
-    let { markup, inline = false }: Props = $props();
+    let { markup, inline = false, note = false }: Props = $props();
 
-    let parsed =
-        $derived(markup instanceof Markup
+    let parsed = $derived(
+        markup instanceof Markup
             ? markup
             : Markup.words(
-                  Array.isArray(markup) ? markup.join('\n\n') : markup
-              ));
+                  Array.isArray(markup) ? markup.join('\n\n') : markup,
+              ),
+    );
 
     let spaces = $derived(parsed.spaces);
 
     // Convert sequences of paragraphs that start with bullets into an HTML list.
-    let paragraphsAndLists = $derived(parsed.paragraphs.reduce(
-        (stuff: ParagraphOrList[], next: Paragraph) => {
-            if (next.isBulleted()) {
-                const previous = stuff.at(-1);
-                if (previous instanceof Paragraph)
-                    return [...stuff, { items: [next] }];
-                else if (previous !== undefined) {
-                    previous.items.push(next);
-                    return stuff;
-                } else return [{ items: [next] }];
-            } else return [...stuff, next];
-        },
-        []
-    ));
+    let paragraphsAndLists = $derived(
+        parsed.paragraphs.reduce(
+            (stuff: ParagraphOrList[], next: Paragraph) => {
+                if (next.isBulleted()) {
+                    const previous = stuff.at(-1);
+                    if (previous instanceof Paragraph)
+                        return [...stuff, { items: [next] }];
+                    else if (previous !== undefined) {
+                        previous.items.push(next);
+                        return stuff;
+                    } else return [{ items: [next] }];
+                } else return [...stuff, next];
+            },
+            [],
+        ),
+    );
 </script>
 
 {#if spaces}
@@ -46,31 +50,38 @@
         {#each parsed.asLine().paragraphs[0].segments as segment}
             <SegmentHTMLView {segment} {spaces} alone={false} />
         {/each}
-    {:else}{#each paragraphsAndLists as paragraphOrList, index}{#if paragraphOrList instanceof Paragraph}
-                <p
-                    class="paragraph"
-                    class:animated={$animationFactor > 0}
-                    style="--delay:{$animationDuration * index * 0.1}ms"
-                    >{#each paragraphOrList.segments as segment}<SegmentHTMLView
-                            {segment}
-                            {spaces}
-                            alone={paragraphOrList.segments.length === 1}
-                        />{/each}</p
-                >{:else}<ul
-                    class:animated={$animationFactor > 0}
-                    style="--delay:{$animationDuration * index * 0.1}ms"
-                    >{#each paragraphOrList.items as paragraph}<li
-                            >{#each paragraph.segments as segment}<SegmentHTMLView
-                                    {segment}
-                                    {spaces}
-                                    alone={paragraph.segments.length === 1}
-                                />{/each}</li
-                        >{/each}</ul
-                >{/if}{/each}
-    {/if}
+    {:else}<div class="markup"
+            >{#each paragraphsAndLists as paragraphOrList, index}{#if paragraphOrList instanceof Paragraph}
+                    <p
+                        class="paragraph"
+                        class:note
+                        class:animated={$animationFactor > 0}
+                        style="--delay:{$animationDuration * index * 0.1}ms"
+                        >{#each paragraphOrList.segments as segment}<SegmentHTMLView
+                                {segment}
+                                {spaces}
+                                alone={paragraphOrList.segments.length === 1}
+                            />{/each}</p
+                    >{:else}<ul
+                        class:animated={$animationFactor > 0}
+                        style="--delay:{$animationDuration * index * 0.1}ms"
+                        >{#each paragraphOrList.items as paragraph}<li
+                                >{#each paragraph.segments as segment}<SegmentHTMLView
+                                        {segment}
+                                        {spaces}
+                                        alone={paragraph.segments.length === 1}
+                                    />{/each}</li
+                            >{/each}</ul
+                    >{/if}{/each}
+        </div>{/if}
 {:else}no spaces{/if}
 
 <style>
+    .markup {
+        display: flex;
+        flex-direction: column;
+        gap: 1em;
+    }
     .paragraph.animated {
         transform: scaleY(0);
         animation-name: pop;
@@ -99,11 +110,17 @@
         margin-inline-start: 0;
     }
 
-    p:last-of-type {
-        margin-block-end: 0em;
+    .note {
+        font-size: var(--wordplay-small-font-size);
     }
 
-    p:not(:global(:first-child)) {
-        margin-block-start: 1em;
+    p,
+    ul {
+        margin-block-start: 0em;
+        margin-block-end: 1em;
+    }
+
+    p:last-child {
+        margin-block-end: 0;
     }
 </style>
