+++ src/components/project/CreatorList.svelte
@@ -13,9 +13,9 @@
 
     interface Props {
         uids: string[];
-        add: (uid: string) => void;
-        remove: (uid: string) => void;
-        removable: (uid: string) => boolean;
+        add?: undefined | ((uid: string) => void);
+        remove?: undefined | ((uid: string) => void);
+        removable?: undefined | ((uid: string) => boolean);
         editable: boolean;
         anonymize: boolean;
     }
@@ -45,7 +45,7 @@
                 unknown = true;
             } else {
                 unknown = false;
-                add(userID);
+                if (add) add(userID);
             }
             emailOrUsername = '';
         }
@@ -62,42 +62,13 @@
     });
 </script>
 
-{#if editable}
-    <form class="form" onsubmit={addCreator}>
-        <TextField
-            bind:text={emailOrUsername}
-            placeholder={$locales.get(
-                (l) => l.ui.dialog.share.field.emailOrUsername.placeholder,
-            )}
-            description={$locales.get(
-                (l) => l.ui.dialog.share.field.emailOrUsername.description,
-            )}
-            validator={validCollaborator}
-        />
-        <Button
-            submit
-            background
-            tip={$locales.get((l) => l.ui.dialog.share.button.submit)}
-            active={validCollaborator(emailOrUsername)}
-            action={() => undefined}>&gt;</Button
-        >
-        {#if adding}<Spinning label="" />{/if}
-        {#if unknown}
-            <Feedback
-                >{$locales.get(
-                    (l) => l.ui.dialog.share.error.unknown,
-                )}</Feedback
-            >{/if}
-    </form>
-{/if}
-
 <div class="people">
     {#each Object.entries(creators) as [uid, creator]}
         <div class="person"
             >{#if creator}<CreatorView
                     {anonymize}
                     {creator}
-                />{:else}?{/if}{#if editable}<Button
+                />{:else}?{/if}{#if editable && removable && remove}<Button
                     tip={$locales.get(
                         (l) => l.ui.project.button.removeCollaborator,
                     )}
@@ -106,25 +77,56 @@
                 >{/if}</div
         >
     {/each}
+    {#if editable}
+        <form class="form" onsubmit={addCreator}>
+            <TextField
+                bind:text={emailOrUsername}
+                placeholder={$locales.get(
+                    (l) => l.ui.dialog.share.field.emailOrUsername.placeholder,
+                )}
+                description={$locales.get(
+                    (l) => l.ui.dialog.share.field.emailOrUsername.description,
+                )}
+                validator={validCollaborator}
+            />
+            <Button
+                submit
+                background
+                padding={false}
+                tip={$locales.get((l) => l.ui.dialog.share.button.submit)}
+                active={validCollaborator(emailOrUsername)}
+                action={() => undefined}>&gt;</Button
+            >
+            {#if adding}<Spinning label="" />{/if}
+            {#if unknown}
+                <Feedback
+                    >{$locales.get(
+                        (l) => l.ui.dialog.share.error.unknown,
+                    )}</Feedback
+                >{/if}
+        </form>
+    {/if}
 </div>
 
 <style>
     .people {
         display: flex;
-        flex-direction: column;
+        flex-direction: row;
+        flex-wrap: wrap;
         gap: var(--wordplay-spacing);
-        margin-block-start: calc(2 * var(--wordplay-spacing));
-        justify-content: center;
+        font-size: var(--wordplay-small-font-size);
     }
 
     .person {
         display: flex;
         flex-direction: row;
-        gap: var(--wordplay-spacing);
+        gap: calc(var(--wordplay-spacing) / 2);
         align-items: center;
     }
 
-    form {
-        margin-block-start: calc(2 * var(--wordplay-spacing));
+    .form {
+        display: flex;
+        flex-direction: row;
+        flex-wrap: nowrap;
     }
 </style>
