+++ src/components/widgets/Dialog.svelte
@@ -46,7 +46,8 @@
 {#if button}
     <Button tip={button.tip} action={() => (show = true)}
         >{#if button.icon}<Emoji>{button.icon}</Emoji>
-        {/if}{button.label}</Button
+        {/if}
+        {button.label}</Button
     >
 {/if}
 <!-- svelte-ignore a11y_no_noninteractive_element_interactions -->
