+++ src/components/widgets/Labeled.svelte
@@ -1,0 +1,27 @@
+<script lang="ts">
+    import { type Snippet } from 'svelte';
+
+    const { label, children }: { label: string; children: Snippet } = $props();
+</script>
+
+<label>
+    <span class="label">{label}</span>
+    {@render children()}
+</label>
+
+<style>
+    label {
+        display: flex;
+        flex-direction: row;
+        gap: var(--wordplay-spacing);
+        align-items: baseline;
+    }
+
+    label .label {
+        font-style: italic;
+        font-size: var(--wordplay-small-font-size);
+        display: flex;
+        flex-direction: row;
+        gap: var(--wordplay-spacing);
+    }
+</style>
