+++ static/locales/example/example.json
@@ -3069,7 +3069,7 @@
                 "palette": "$?",
                 "docs": "$?",
                 "source": "$?",
-                "chat": "$?"
+                "collaborate": "$?"
             },
             "button": {
                 "collapse": "$?"
@@ -3723,30 +3723,6 @@
         "template": {
             "unwritten": "$?",
             "unparsable": "$?"
-        },
-        "chat": {
-            "label": "$?",
-            "field": {
-                "message": {
-                    "description": "$?",
-                    "placeholder": "$?"
-                }
-            },
-            "button": {
-                "submit": {
-                    "label": "$?",
-                    "tip": "$?"
-                },
-                "start": {
-                    "label": "$?",
-                    "tip": "$?"
-                }
-            },
-            "error": {
-                "unowned": "$?",
-                "empty": "$?"
-            },
-            "prompt": "$?"
         }
     },
     "moderation": {
