+++ static/locales/hi-IN/hi-IN.json
@@ -4064,8 +4064,7 @@
                 "output": "$~अवस्था",
                 "palette": "$~पैलेट",
                 "docs": "$~मार्गदर्शक",
-                "source": "$~स्रोत",
-                "chat": "$~बात करना"
+                "source": "$~स्रोत"
             },
             "button": {
                 "collapse": "$~विंडो को छोटा करें"
@@ -4812,30 +4811,6 @@
         "template": {
             "unwritten": "$~टीबीडी",
             "unparsable": "$~अनपर्सेबल टेम्पलेट: $1"
-        },
-        "chat": {
-            "label": "$~बात करना",
-            "field": {
-                "message": {
-                    "description": "$~सबमिट करने के लिए चैट संदेश",
-                    "placeholder": "$~संदेश लिखें"
-                }
-            },
-            "button": {
-                "submit": {
-                    "label": "$~भेजना",
-                    "tip": "$~अपने सहयोगियों को संदेश भेजें"
-                },
-                "start": {
-                    "label": "$~चैटिंग शुरू करें",
-                    "tip": "$~अपने आप से या दूसरों से बातचीत शुरू करें।"
-                }
-            },
-            "error": {
-                "unowned": "$~इस प्रोजेक्ट का कोई स्वामी नहीं है, इसलिए इस पर चैट नहीं हो सकती।",
-                "empty": "$~कोई संदेश नहीं."
-            },
-            "prompt": "$~क्या आप नोट्स लेना चाहते हैं या दूसरों के साथ सहयोग करना चाहते हैं?"
         }
     },
     "moderation": {
