+++ static/locales/ja-JP/ja-JP.json
@@ -4064,8 +4064,7 @@
                 "output": "$~ステージ",
                 "palette": "$~パレット",
                 "docs": "$~ガイド",
-                "source": "$~ソース",
-                "chat": "$~チャット"
+                "source": "$~ソース"
             },
             "button": {
                 "collapse": "$~ウィンドウを折りたたむ"
@@ -4807,30 +4806,6 @@
         "template": {
             "unwritten": "$~未定",
             "unparsable": "$~解析できないテンプレート: $1"
-        },
-        "chat": {
-            "label": "$~チャット",
-            "field": {
-                "message": {
-                    "description": "$~送信するチャットメッセージ",
-                    "placeholder": "$~メッセージを入力してください"
-                }
-            },
-            "button": {
-                "submit": {
-                    "label": "$~送信",
-                    "tip": "$~共同作業者にメッセージを送信する"
-                },
-                "start": {
-                    "label": "$~チャットを始める",
-                    "tip": "$~自分自身または他の人とチャットを開始します。"
-                }
-            },
-            "error": {
-                "unowned": "$~このプロジェクトには所有者がいないため、チャットを行うことはできません。",
-                "empty": "$~メッセージはありません。"
-            },
-            "prompt": "$~メモを取ったり、他の人と共同作業したりしたいですか?"
         }
     },
     "moderation": {
